+++ src/graphsenselib/ingest/cli.py
@@ -1,3 +1,6 @@
+import logging
+import sys
+
 import click
 
 from ..cli.common import require_currency, require_environment
@@ -7,6 +10,8 @@ from ..schema import GraphsenseSchemas
 from .common import INGEST_SINKS
 from .factory import IngestFactory
 from .parquet import SCHEMA_MAPPING
+
+logger = logging.getLogger(__name__)
 
 
 @click.group()
@@ -55,7 +60,7 @@ def ingesting():
     type=int,
     required=False,
     default=3600,
-    help="end block (default: last available block)",
+    help="Web3 API timeout in seconds (default: 3600s)",
 )
 @click.option(
     "--info",
@@ -95,7 +100,9 @@ def ingest(
     """
     ks_config = config.get_keyspace_config(env, currency)
     provider = ks_config.ingest_config.node_reference
-    parquet_file_sink = ks_config.ingest_config.raw_keyspace_file_sink_directory
+    parquet_file_sink = ks_config.ingest_config.raw_keyspace_file_sinks.get(
+        "parquet", None
+    ).directory
 
     if create_schema:
         GraphsenseSchemas().create_keyspace_if_not_exist(
@@ -122,6 +129,124 @@ def ingest(
             user_start_block=start_block,
             user_end_block=end_block,
             batch_size=batch_size,
+            info=info,
+            previous_day=previous_day,
+            provider_timeout=timeout,
+        )
+
+
+@ingesting.command("to-csv")
+@require_environment()
+@require_currency(required=True)
+@click.option(
+    "--start-block",
+    type=int,
+    required=False,
+    help="start block (default: 0)",
+)
+@click.option(
+    "--end-block",
+    type=int,
+    required=False,
+    help="end block (default: last available block)",
+)
+@click.option(
+    "--continue",
+    "continue_export",
+    is_flag=True,
+    help="continue from export from position",
+)
+@click.option(
+    "--batch-size",
+    type=int,
+    default=10,
+    help="number of blocks to export (write) at a time (default: 10)",
+)
+@click.option(
+    "--file-batch-size",
+    type=int,
+    default=1000,
+    help="number of blocks to export to a CSV file (default: 1000)",
+)
+@click.option(
+    "--partition-batch-size",
+    type=int,
+    default=1_000_000,
+    help="number of blocks to export in partition (default: 1_000_000)",
+)
+@click.option(
+    "-p",
+    "--previous_day",
+    is_flag=True,
+    help="only ingest blocks up to the previous day, "
+    "since currency exchange rates might not be "
+    "available for the current day",
+)
+@click.option(
+    "--info",
+    is_flag=True,
+    help="display block information and exit",
+)
+@click.option(
+    "--timeout",
+    type=int,
+    required=False,
+    default=3600,
+    help="Web3 API timeout in seconds (default: 3600s)",
+)
+def export_csv(
+    env,
+    currency,
+    start_block,
+    end_block,
+    continue_export,
+    batch_size,
+    file_batch_size,
+    partition_batch_size,
+    previous_day,
+    info,
+    timeout,
+):
+    """Exports raw cryptocurrency data to gziped csv files.
+    \f
+    Args:
+        env (str): Environment to work on
+        currency (str): currency to work on
+    """
+
+    if currency != "eth":
+        logger.error("Csv export is only supported for eth at the moment.")
+        sys.exit(1)
+
+    ks_config = config.get_keyspace_config(env, currency)
+    provider = ks_config.ingest_config.node_reference
+    csv_directory_config = ks_config.ingest_config.raw_keyspace_file_sinks.get(
+        "csv", None
+    )
+
+    if csv_directory_config is None:
+        logger.error(
+            "Please provide an output directory in your "
+            "config (raw_keyspace_file_sinks.csv.directory)"
+        )
+        sys.exit(1)
+
+    csv_directory = csv_directory_config.directory
+
+    from .account import export_csv
+
+    with DbFactory().from_config(env, currency) as db:
+        export_csv(
+            db=db,
+            currency=currency,
+            provider_uri=provider,
+            directory=csv_directory,
+            user_start_block=start_block,
+            user_end_block=end_block,
+            continue_export=continue_export,
+            batch_size=batch_size,
+            file_batch_size=file_batch_size,
+            partition_batch_size=partition_batch_size,
             info=info,
             previous_day=previous_day,
             provider_timeout=timeout,